modified for 64bit

+++ cxl_mailbox.h
@@ -162,12 +162,12 @@ enum mailbox_ret_codes{
 };
 
 void print_config_header(struct pci_dev *pdev);
-void cxl_mailbox_get_timestamp(uint32_t mailbox_base_address);
+void cxl_mailbox_get_timestamp(uint64_t mailbox_base_address);
 void print_extended_config(struct pci_dev *pdev);
 uint16_t get_dvsec_register_locator_offset(struct pci_dev *pdev);
-uint32_t get_mailbox_base_address (struct pci_dev *pdev);
+uint64_t get_mailbox_base_address (struct pci_dev *pdev);
 uint32_t get_register_block_number_from_header(registerLocator *register_locator);
-int send_mailbox_command(uint32_t mailbox_base_address, uint16_t command,uint16_t *payload_size, uint32_t *payload, uint16_t *ret_code);
+int send_mailbox_command(uint64_t mailbox_base_address, uint16_t command,uint16_t *payload_size, uint32_t *payload, uint16_t *ret_code);
 bool check_mailbox_ready(mailbox_registers *mb_regs);
 void mailbox_write_payload(mailbox_registers *mb_regs, uint16_t payload_length, uint32_t *payload);
 void read_payload(mailbox_registers *mb_regs, uint16_t payload_length, uint32_t *payload);
